Simplify post-1.5.0 hotspots without new abstraction layers.
Pull question-quality under 1k lines, share delete/batch/cost helpers, fix history hydrate collisions, and keep HTTP out of record-attempt.
