+++ src/lib/users/delete-app-data.server.ts
@@ -1,7 +1,5 @@
 import { connectDb } from '$lib/server/db';
-import { UserProfile } from '$lib/users/model.server';
-import { Referral } from '$lib/referrals/model.server';
-import { FrqAttempt } from '$lib/frq/model.server';
+import { deleteAppDataDocuments } from '$lib/users/delete-app-data-documents.server';
 
 /**
  * Deletes app-owned rows for one or more users (profile, FRQ attempts, referrals).
@@ -12,14 +10,5 @@ export async function deleteAppDataForUsers(userIds: string | string[]): Promise
 	if (ids.length === 0) return;
 
 	await connectDb();
-
-	const userIdFilter = ids.length === 1 ? ids[0]! : { $in: ids };
-
-	await Promise.all([
-		UserProfile.deleteMany({ userId: userIdFilter }),
-		FrqAttempt.deleteMany({ userId: userIdFilter }),
-		Referral.deleteMany({
-			$or: [{ referrerUserId: userIdFilter }, { referredUserId: userIdFilter }]
-		})
-	]);
+	await deleteAppDataDocuments(ids);
 }
