+++ tests/unit/users/practice-history.test.ts
@@ -1,11 +1,14 @@
 import { describe, expect, it, vi } from 'vitest';
 
-const mocks = vi.hoisted(() => ({ find: vi.fn() }));
+const mocks = vi.hoisted(() => ({ find: vi.fn(), getQuestionsLookupMap: vi.fn() }));
 
 vi.mock('$lib/server/db', () => ({ connectDb: vi.fn() }));
 vi.mock('$lib/frq/model.server', () => ({ FrqAttempt: { find: mocks.find } }));
+vi.mock('$lib/questions/storage.server', () => ({
+	getQuestionsLookupMap: mocks.getQuestionsLookupMap
+}));
 
-import { getPracticeHistoryPage } from '$lib/users/history.server';
+import { getPracticeHistoryPage, hydratePracticeHistoryItems } from '$lib/users/history.server';
 
 describe('getPracticeHistoryPage', () => {
 	it('merges authenticated FRQ attempts with MCQ history and sorts them together', async () => {
@@ -44,5 +47,56 @@ describe('getPracticeHistoryPage', () => {
 		expect(page.total).toBe(2);
 		expect(page.items.map((item) => item.kind)).toEqual(['frq', 'mcq']);
 		expect(page.items[0]?.attempt.questionId).toBe('frq-question-1');
+	});
+});
+
+describe('hydratePracticeHistoryItems', () => {
+	it('keeps distinct MCQ attempts that share the same questionId', async () => {
+		mocks.getQuestionsLookupMap.mockResolvedValue(
+			new Map([
+				[
+					'shared-question',
+					{
+						id: 'shared-question',
+						stem: 'Shared question',
+						choices: [],
+						correctAnswer: 'A'
+					}
+				]
+			])
+		);
+
+		const hydrated = await hydratePracticeHistoryItems([
+			{
+				kind: 'mcq',
+				attempt: {
+					questionId: 'shared-question',
+					apClass: 'AP Biology',
+					unit: 'Unit 1',
+					wasCorrect: true,
+					selectedAnswer: 'A',
+					attemptedAt: new Date('2026-07-01T00:00:00.000Z')
+				},
+				question: null
+			},
+			{
+				kind: 'mcq',
+				attempt: {
+					questionId: 'shared-question',
+					apClass: 'AP Biology',
+					unit: 'Unit 1',
+					wasCorrect: false,
+					selectedAnswer: 'B',
+					attemptedAt: new Date('2026-07-02T00:00:00.000Z')
+				},
+				question: null
+			}
+		]);
+
+		expect(hydrated).toHaveLength(2);
+		expect(hydrated[0]?.attempt.wasCorrect).toBe(true);
+		expect(hydrated[1]?.attempt.wasCorrect).toBe(false);
+		expect(hydrated[0]?.question?.id).toBe('shared-question');
+		expect(hydrated[1]?.question?.id).toBe('shared-question');
 	});
 });
